fix bugs

+++ src/main/java/com/longforus/kotlincodesorter/sort/Sorter.kt
@@ -19,17 +19,21 @@ class Sorter
 
     fun realSort(classOrObject: KtClassOrObject) {
         val declarations = classOrObject.declarations
+        val before = declarations.hashCode()
         val sort = CommonSortStrategy(declarations).sort()
+        val after = sort.hashCode()
         sort.forEach {
             if (it is KtClassOrObject) {
                 realSort(it)
             }
         }
-        sort.forEach {
-            classOrObject.addDeclaration(it)
-        }
-        declarations.forEach {
-            it.delete()
+        if (before != after) {
+            sort.forEach {
+                classOrObject.addDeclaration(it)
+            }
+            declarations.forEach {
+                it.delete()
+            }
         }
     }
